Permissions checking for file operations

+++ lib/src/main/java/edu/utexas/tacc/tapis/files/lib/services/FileOpsService.java
@@ -14,6 +14,7 @@ import org.slf4j.LoggerFactory;
 import org.jetbrains.annotations.NotNull;
 
 import javax.inject.Inject;
+import javax.ws.rs.NotAuthorizedException;
 import javax.ws.rs.NotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
@@ -28,29 +29,32 @@ public class FileOpsService implements IFileOpsService {
 
     private static final Logger log = LoggerFactory.getLogger(FileOpsService.class);
     private static final int MAX_LISTING_SIZE = Constants.MAX_LISTING_SIZE;
-    private final TenantManager tenantManager;
+    private final FilePermsService permsService;
 
     // 0 = tenantBaseUri, 1=systemId, 2=path
     private String TAPIS_FILES_URI_FORMAT = "tapis://{0}/{1}/{2}";
 
     @Inject
-    public FileOpsService(TenantManager tenantManager) {
-        this.tenantManager = tenantManager;
+    public FileOpsService(FilePermsService permsService) {
+        this.permsService = permsService;
     }
 
 
     @Override
-    public List<FileInfo> ls(IRemoteDataClient client, @NotNull String path) throws ServiceException, NotFoundException {
+    public List<FileInfo> ls(IRemoteDataClient client, @NotNull String path) throws ServiceException, NotFoundException, NotAuthorizedException {
         return ls(client, path, MAX_LISTING_SIZE, 0);
     }
 
     @Override
     public List<FileInfo> ls(IRemoteDataClient client, @NotNull String path, long limit, long offset) throws ServiceException, NotFoundException {
         try {
+            String cleanedPath = FilenameUtils.normalize(path);
+            boolean permitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), cleanedPath, FileInfo.Permission.MODIFY);
+            if (!permitted) {
+                String msg = Utils.getMsg("FILES_NOT_AUTHORIZED", client.getOboTenant(), client.getOboUser(), client.getSystemId(), path);
+                throw new NotAuthorizedException(msg);
+            }
             List<FileInfo> listing = client.ls(path, limit, offset);
-            listing.forEach(f-> {
-
-            });
             return listing;
         } catch (IOException ex) {
             String msg = Utils.getMsg("FILES_OPSC_ERR", client.getOboTenant(), client.getOboUser(), "listing",
@@ -61,9 +65,14 @@ public class FileOpsService implements IFileOpsService {
     }
 
     @Override
-    public void mkdir(IRemoteDataClient client, String path) throws ServiceException {
+    public void mkdir(IRemoteDataClient client, String path) throws ServiceException, NotAuthorizedException {
         try {
             String cleanedPath = FilenameUtils.normalize(path);
+            boolean permitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), cleanedPath, FileInfo.Permission.MODIFY);
+            if (!permitted) {
+                String msg = Utils.getMsg("FILES_NOT_AUTHORIZED", client.getOboTenant(), client.getOboUser(), client.getSystemId(), path);
+                throw new NotAuthorizedException(msg);
+            }
             client.mkdir(cleanedPath);
         } catch (IOException ex) {
             String msg = Utils.getMsg("FILES_OPSC_ERR", client.getOboTenant(), client.getOboUser(), "mkdir",
@@ -74,8 +83,13 @@ public class FileOpsService implements IFileOpsService {
     }
 
     @Override
-    public void insert(IRemoteDataClient client, String path, @NotNull InputStream inputStream) throws ServiceException {
+    public void insert(IRemoteDataClient client, String path, @NotNull InputStream inputStream) throws ServiceException, NotFoundException, NotAuthorizedException {
         try {
+            boolean permitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), path, FileInfo.Permission.MODIFY);
+            if (!permitted) {
+                String msg = Utils.getMsg("FILES_NOT_AUTHORIZED", client.getOboTenant(), client.getOboUser(), client.getSystemId(), path);
+                throw new NotAuthorizedException(msg);
+            }
             client.insert(path, inputStream);
         } catch (IOException ex) {
             String msg = Utils.getMsg("FILES_OPSC_ERR", client.getOboTenant(), client.getOboUser(), "insert",
@@ -86,8 +100,14 @@ public class FileOpsService implements IFileOpsService {
     }
 
     @Override
-    public void move(IRemoteDataClient client, String path, String newPath) throws ServiceException, NotFoundException {
+    public void move(IRemoteDataClient client, String path, String newPath) throws ServiceException, NotFoundException, NotAuthorizedException {
         try {
+            boolean srcPermitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), path, FileInfo.Permission.MODIFY);
+            boolean destPermitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), newPath, FileInfo.Permission.MODIFY);
+            if (!srcPermitted || !destPermitted) {
+                String msg = Utils.getMsg("FILES_NOT_AUTHORIZED", client.getOboTenant(), client.getOboUser(), client.getSystemId(), path);
+                throw new NotAuthorizedException(msg);
+            }
             client.move(path, newPath);
         } catch (IOException ex) {
             String msg = Utils.getMsg("FILES_OPSC_ERR", client.getOboTenant(), client.getOboUser(), "move",
@@ -98,8 +118,14 @@ public class FileOpsService implements IFileOpsService {
     }
 
     @Override
-    public void copy(IRemoteDataClient client, String path, String newPath) throws ServiceException, NotFoundException {
+    public void copy(IRemoteDataClient client, String path, String newPath) throws ServiceException, NotFoundException, NotAuthorizedException {
         try {
+            boolean srcPermitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), path, FileInfo.Permission.READ);
+            boolean destPermitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), newPath, FileInfo.Permission.MODIFY);
+            if (!srcPermitted || !destPermitted) {
+                String msg = Utils.getMsg("FILES_NOT_AUTHORIZED", client.getOboTenant(), client.getOboUser(), client.getSystemId(), path);
+                throw new NotAuthorizedException(msg);
+            }
             client.copy(path, newPath);
         } catch (IOException ex) {
             String msg = Utils.getMsg("FILES_OPSC_ERR", client.getOboTenant(), client.getOboUser(), "copy",
@@ -111,8 +137,13 @@ public class FileOpsService implements IFileOpsService {
 
 
     @Override
-    public void delete(IRemoteDataClient client, @NotNull String path) throws ServiceException, NotFoundException {
+    public void delete(IRemoteDataClient client, @NotNull String path) throws ServiceException, NotFoundException, NotAuthorizedException {
         try {
+            boolean permitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), path, FileInfo.Permission.MODIFY);
+            if (!permitted) {
+                String msg = Utils.getMsg("FILES_NOT_AUTHORIZED", client.getOboTenant(), client.getOboUser(), client.getSystemId(), path);
+                throw new NotAuthorizedException(msg);
+            }
             client.delete(path);
         } catch (IOException ex) {
             String msg = Utils.getMsg("FILES_OPSC_ERR", client.getOboTenant(), client.getOboUser(), "delete",
@@ -132,11 +163,14 @@ public class FileOpsService implements IFileOpsService {
      * @throws ServiceException
      */
     @Override
-    public InputStream getStream(IRemoteDataClient client, String path) throws ServiceException, NotFoundException {
-        // Try with resources to auto close the stream
-        // Pushing the InputStream to a bufferedInputStream is memory efficient
-        // way to auto close the initial input stream.
+    public InputStream getStream(IRemoteDataClient client, String path) throws ServiceException, NotFoundException, NotAuthorizedException{
+
         try {
+            boolean permitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), path, FileInfo.Permission.READ);
+            if (!permitted) {
+                String msg = Utils.getMsg("FILES_NOT_AUTHORIZED", client.getOboTenant(), client.getOboUser(), client.getSystemId(), path);
+                throw new NotAuthorizedException(msg);
+            }
             InputStream fileStream = client.getStream(path);
             return fileStream;
         } catch (IOException ex) {
@@ -148,8 +182,13 @@ public class FileOpsService implements IFileOpsService {
     }
 
     @Override
-    public InputStream getBytes(IRemoteDataClient client, @NotNull String path, long startByte, long count) throws ServiceException, NotFoundException {
+    public InputStream getBytes(IRemoteDataClient client, @NotNull String path, long startByte, long count) throws ServiceException, NotFoundException, NotAuthorizedException {
         try  {
+            boolean permitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), path, FileInfo.Permission.READ);
+            if (!permitted) {
+                String msg = Utils.getMsg("FILES_NOT_AUTHORIZED", client.getOboTenant(), client.getOboUser(), client.getSystemId(), path);
+                throw new NotAuthorizedException(msg);
+            }
             InputStream fileStream = client.getBytesByRange(path, startByte, count);
             return fileStream;
         } catch (IOException ex) {
@@ -161,9 +200,14 @@ public class FileOpsService implements IFileOpsService {
     }
 
     @Override
-    public InputStream more(IRemoteDataClient client, @NotNull String path, long startPage) throws ServiceException, NotFoundException {
+    public InputStream more(IRemoteDataClient client, @NotNull String path, long startPage) throws ServiceException, NotFoundException, NotAuthorizedException{
         long startByte = (startPage - 1) * 1024;
         try  {
+            boolean permitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), path, FileInfo.Permission.READ);
+            if (!permitted) {
+                String msg = Utils.getMsg("FILES_NOT_AUTHORIZED", client.getOboTenant(), client.getOboUser(), client.getSystemId(), path);
+                throw new NotAuthorizedException(msg);
+            }
             InputStream fileStream = client.getBytesByRange(path, startByte, startByte + 1023);
             return fileStream;
         } catch (IOException ex) {
@@ -182,10 +226,15 @@ public class FileOpsService implements IFileOpsService {
      * @throws IOException
      */
     @Override
-    public void getZip(IRemoteDataClient client, @NotNull OutputStream outputStream, @NotNull String path) throws ServiceException {
+    public void getZip(IRemoteDataClient client, @NotNull OutputStream outputStream, @NotNull String path) throws ServiceException, NotAuthorizedException {
+
+        boolean permitted = permsService.isPermitted(client.getOboTenant(), client.getOboUser(), client.getSystemId(), path, FileInfo.Permission.READ);
+        if (!permitted) {
+            String msg = Utils.getMsg("FILES_NOT_AUTHORIZED", client.getOboTenant(), client.getOboUser(), client.getSystemId(), path);
+            throw new NotAuthorizedException(msg);
+        }
 
         //TODO: This should be made for recursive listings
-
         List<FileInfo> listing = this.ls(client, path);
         try (ZipOutputStream zipStream = new ZipOutputStream(outputStream)) {
             for (FileInfo fileInfo: listing) {
